+++ mathics/builtin/no_meaning/__init__.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+"""
+Operators without Built-in Meanings
+
+Not all operators recognized by the Mathics3 are associated with functions that have built‐in meanings.
+You can use these operators as a way to build up your own notation within Mathics3.
+"""
+
+# This tells documentation how to sort this module
+sort_order = "mathics.builtin.operators-without-built-in-meanings"
